EZP-24800 Support allowed classes limitation for object relation fields
