+++ lib/FieldType/Mapper/RelationFormMapper.php
@@ -9,12 +9,12 @@
 namespace EzSystems\RepositoryForms\FieldType\Mapper;
 
 use EzSystems\RepositoryForms\Data\FieldDefinitionData;
-use EzSystems\RepositoryForms\FieldType\FieldDefinitionFormMapperInterface;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class RelationFormMapper implements FieldDefinitionFormMapperInterface
+class RelationFormMapper extends AbstractRelationFormMapper
 {
     public function mapFieldDefinitionForm(FormInterface $fieldDefinitionForm, FieldDefinitionData $data)
     {
@@ -23,11 +23,21 @@ class RelationFormMapper implements FieldDefinitionFormMapperInterface
                 'required' => false,
                 'property_path' => 'fieldSettings[selectionRoot]',
                 'label' => 'field_definition.ezobjectrelation.selection_root',
+            ])
+            ->add('selectionContentTypes', ChoiceType::class, [
+                'choices' => $this->getContentTypeHash(),
+                'choices_as_values' => true,
+                'expanded' => false,
+                'multiple' => true,
+                'required' => false,
+                'property_path' => 'fieldSettings[selectionContentTypes]',
+                'label' => 'field_definition.ezobjectrelation.selection_content_types',
             ]);
     }
 
     /**
      * Fake method to set the translation domain for the extractor.
+     * @param OptionsResolver $resolver
      */
     public function configureOptions(OptionsResolver $resolver)
     {
