+++ lib/FieldType/Mapper/RelationListFormMapper.php
@@ -8,49 +8,16 @@
  */
 namespace EzSystems\RepositoryForms\FieldType\Mapper;
 
-use eZ\Publish\API\Repository\ContentTypeService;
-use eZ\Publish\Core\FieldType\RelationList\Type;
-use eZ\Publish\Core\Helper\TranslationHelper;
 use EzSystems\RepositoryForms\Data\FieldDefinitionData;
-use EzSystems\RepositoryForms\FieldType\FieldDefinitionFormMapperInterface;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class RelationListFormMapper implements FieldDefinitionFormMapperInterface
+class RelationListFormMapper extends AbstractRelationFormMapper
 {
-    /**
-     * @var ContentTypeService Used to fetch list of available content types
-     */
-    protected $contentTypeService;
-
-    /**
-     * @var TranslationHelper Translation helper, for translated content type names
-     */
-    protected $translationHelper;
-
-    /**
-     * @param ContentTypeService $contentTypeService
-     * @param TranslationHelper $translationHelper
-     */
-    public function __construct(ContentTypeService $contentTypeService, TranslationHelper $translationHelper)
-    {
-        $this->contentTypeService = $contentTypeService;
-        $this->translationHelper = $translationHelper;
-    }
-
     public function mapFieldDefinitionForm(FormInterface $fieldDefinitionForm, FieldDefinitionData $data)
     {
-        // Fill a hash with all content types and their ids
-        $contentTypeHash = [];
-        foreach ($this->contentTypeService->loadContentTypeGroups() as $contentTypeGroup) {
-            foreach ($this->contentTypeService->loadContentTypes($contentTypeGroup) as $contentType) {
-                $contentTypeHash[$this->translationHelper->getTranslatedByProperty($contentType, 'names')] = $contentType->identifier;
-            }
-        }
-        ksort($contentTypeHash);
-
         $fieldDefinitionForm
             ->add('selectionDefaultLocation', HiddenType::class, [
                 'required' => false,
@@ -58,7 +25,7 @@ class RelationListFormMapper implements FieldDefinitionFormMapperInterface
                 'label' => 'field_definition.ezobjectrelationlist.selection_default_location',
             ])
             ->add('selectionContentTypes', ChoiceType::class, [
-                'choices' => $contentTypeHash,
+                'choices' => $this->getContentTypeHash(),
                 'choices_as_values' => true,
                 'expanded' => false,
                 'multiple' => true,
@@ -70,6 +37,7 @@ class RelationListFormMapper implements FieldDefinitionFormMapperInterface
 
     /**
      * Fake method to set the translation domain for the extractor.
+     * @param OptionsResolver $resolver
      */
     public function configureOptions(OptionsResolver $resolver)
     {
